requirements.txt
mysql-connector-python
pandas
python-dotenv
python-telegram-bot=13.3